TEMP: all-modules.mnfs: only use SW-912 protocols
We're using all-modules.mnfs as the manifest while hacking on SW-912.

Benoit reports that if too many (more than 16) CPorts are in use in
total, the system breaks down. To prevent that from happening when
activating another bridge (which we need for T1', T2', ..., T5'), remove
anything that isn't needed for SW-912 work from that manifest.

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>